requirements.txt
torch
pyyaml
tiktoken
datasets
pandas
psutil
wandb
lm-eval
